yuggoth: ensure that agenix is mounted at boot
agenix needs the VM's host key in order to decrypt any secrets granted
directly to the VM. the host key is passed into the VM with a mount. the
mount needs to have "neededForBoot" set on order to be available to the
system activation script.

This doesn't work if the mount is under /run, so instead let's put it
under /mnt.