nixos/lan/yuggoth: specify dependencies between VMs, and a startup delay
some VMs depend on other VMs (e.g. paperless depends on db).

a systemd "Requires=" dependency encodes this, but it's not sufficient -
microvm.nix does not detect when the VM has actually booted, so we add
an explicit startup delay too.

this allows essential VMs to come up faster, and reduces load on the
system when it first boots.